+++ ecmascript/compiler/codegen/maple/maple_be/include/cg/x86_64/assembler/asm_assembler.h
@@ -356,7 +356,7 @@ public:
     void Sqrtsd_r(Reg srcReg, Reg destReg) override;
     /* end of X64 instructions */
     /* process stackmap */
-    void RecordStackmap(const std::vector<uint64> &referenceMap, const std::vector<uint64> &deoptInfo) override {}
+    void RecordStackmap(const std::vector<uint8> &referenceMap, const std::vector<uint8> &deoptInfo) override {}
     uint32 GetCurModulePC() override { return 0; }
     void SetLastModulePC(uint32 pc) override {}
